+++ packages/cargo/jns42-generator/src/generators/interiors_rs.rs
