+++ packages/cargo/jns42-generator/src/generators/mod.rs
@@ -2,7 +2,7 @@ pub mod banner;
 pub mod cargo_toml;
 pub mod errors_rs;
 pub mod file;
-pub mod inner_types_rs;
+pub mod interiors_rs;
 pub mod lib_rs;
-pub mod new_types_rs;
 pub mod package;
+pub mod types_rs;
